Separate System and Server tests on the backend and implement github action for testing without the DB dependency (#754)
* Chore: implement server-side tests to the workspace (by adding a db and a server jobs)
* Chore: fix script and add steps
* Chore: separate system and server testing and improve github backend testing script
* Chore: add package-lock.json to git and update comments on main.yml
Did not need to update .gitignore
