+++ server/package.json
@@ -13,6 +13,8 @@
     "lint": "eslint --fix src/**/*.js",
     "test": "node_modules/.bin/mocha --recursive --sort --bail --require @babel/register **/*.spec.js",
     "test:cov": "nyc node_modules/.bin/mocha --recursive --sort --bail --require @babel/register **/*.spec.js",
+    "test:server": "node_modules/.bin/mocha --recursive --sort --bail --require @babel/register **/*.server.spec.js",
+    "test:system": "node_modules/.bin/mocha --recursive --sort --bail --require @babel/register **/*.system.spec.js",
     "start-docker": "node index.js",
     "start-docker-dev": "nodemon -L  node index.js",
     "start-prod": "NODE_ENV=production node index.js",
